// src/models/organisationGridType.ts
// region header
//--------------------------------------------------------------------------------------------------
//
//  Description: Organisation Grid Type
//
//  Copyright:    © 2025 Idox Software Limited.
//
//--------------------------------------------------------------------------------------------------
//
//  Modification History:
//
//  Version Date     Modifier             Issue# Description
// region Version 1.0.0.0
//    001   12.03.25 Sean Flook          GMSCM-1 Initial Revision.
// endregion Version 1.0.0.0
//
//--------------------------------------------------------------------------------------------------
// endregion header

export default interface OrganisationGridType {
  id: number;
  organisation: string;
}